fetch user image inside Avatar component
